add the function for make_set_train into maketest. removed maketrain. call with same flags
